requirements.txt
flask==2.3
gunicorn==21.2
jinja2==3.1
python-dotenv==1.0
requests==2.31
supabase==2.0
tzdata==2024.1
sentry-sdk==2.0